+++ detector.go
@@ -21,25 +21,6 @@ const (
 
 var ErrFail = errors.New("no buildpacks participating")
 
-type Buildpack struct {
-	ID       string `toml:"id" json:"id"`
-	Version  string `toml:"version" json:"version"`
-	Optional bool   `toml:"optional,omitempty" json:"optional,omitempty"`
-}
-
-func (bp Buildpack) dir() string {
-	return escapeID(bp.ID)
-}
-
-func (bp Buildpack) String() string {
-	return bp.ID + "@" + bp.Version
-}
-
-func (bp Buildpack) noOpt() Buildpack {
-	bp.Optional = false
-	return bp
-}
-
 type BuildPlan struct {
 	Entries []BuildPlanEntry `toml:"entries"`
 }
@@ -78,20 +59,6 @@ type DetectConfig struct {
 	runs          *sync.Map
 	UID           int
 	GID           int
-}
-
-func (bp Buildpack) lookup(buildpacksDir string) (*buildpackTOML, error) {
-	bpTOML := buildpackTOML{}
-	bpPath, err := filepath.Abs(filepath.Join(buildpacksDir, bp.dir(), bp.Version))
-	if err != nil {
-		return nil, err
-	}
-	tomlPath := filepath.Join(bpPath, "buildpack.toml")
-	if _, err := toml.DecodeFile(tomlPath, &bpTOML); err != nil {
-		return nil, err
-	}
-	bpTOML.Path = bpPath
-	return &bpTOML, nil
 }
 
 func (c *DetectConfig) process(done []Buildpack) ([]Buildpack, []BuildPlanEntry, error) {
